migration updates

+++ EPS/Modules/Billing/Building.cs
@@ -18,24 +18,15 @@ namespace Modules.Billing
     public class Building : FormClass
     {
         
-
         public Building(frmBilling Form) : base(Form)
         { }
-
-        private List<string> sPermitListSave;
-
-        public List<string> PermitListSave
-        {
-            get { return sPermitListSave; }
-            set { sPermitListSave = value; }
-        }
-
 
         public override void FormLoad()
         {
             RecordFrm.grpAddFees.Visible = true;
         }
 
+       
 
         public override void DisplayAssessmentData()
         {
@@ -74,40 +65,34 @@ namespace Modules.Billing
             string sQuery = string.Empty;
             var db = new EPSConnection(dbConn);
 
-
             try
             {
                 if (MessageBox.Show("Save Billing?", " ", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
                 {
-                    //if (!string.IsNullOrEmpty(RecordFrm.txtBillNo.Text.ToString()))
-                    //{
+                    if (!string.IsNullOrEmpty(RecordFrm.txtBillNo.Text.ToString()))
+                    {
+                        sQuery = $"insert into billing_hist select * from billing where arn = '{RecordFrm.m_sAN}'";
+                        db.Database.ExecuteSqlCommand(sQuery);
 
-                    //    //for (int row = 0; row < RcrdClass.PermitList.Count; row++)
-                    //    //{
+                        sQuery = $"delete from billing where arn = '{RecordFrm.m_sAN}'";
+                        db.Database.ExecuteSqlCommand(sQuery);
 
-                    //    //}
-                    sQuery = $"insert into billing_hist select * from billing where arn = '{RecordFrm.m_sAN}'";
-                    db.Database.ExecuteSqlCommand(sQuery);
+                        sQuery = $"insert into taxdues_hist select * from taxdues where arn = '{RecordFrm.m_sAN}'";
+                        db.Database.ExecuteSqlCommand(sQuery);
 
-                    sQuery = $"delete from billing where arn = '{RecordFrm.m_sAN}'";
-                    db.Database.ExecuteSqlCommand(sQuery);
+                        sQuery = $"delete from taxdues where arn = '{RecordFrm.m_sAN}'";
+                        db.Database.ExecuteSqlCommand(sQuery);
 
-                    sQuery = $"insert into taxdues_hist select * from taxdues where arn = '{RecordFrm.m_sAN}'";
-                    db.Database.ExecuteSqlCommand(sQuery);
+                        sQuery = $"insert into tax_details_hist select * from tax_details where arn = '{RecordFrm.m_sAN}'";
+                        db.Database.ExecuteSqlCommand(sQuery);
 
-                    sQuery = $"delete from taxdues where arn = '{RecordFrm.m_sAN}'";
-                    db.Database.ExecuteSqlCommand(sQuery);
-
-                    sQuery = $"insert into tax_details_hist select * from tax_details where arn = '{RecordFrm.m_sAN}'";
-                    db.Database.ExecuteSqlCommand(sQuery);
-
-                    sQuery = $"delete from tax_details where arn = '{RecordFrm.m_sAN}'";
-                    db.Database.ExecuteSqlCommand(sQuery);
-                }
-                else
-                {
-                    RecordFrm.txtBillNo.Text = GenerateBillNo();
-                }
+                        sQuery = $"delete from tax_details where arn = '{RecordFrm.m_sAN}'";
+                        db.Database.ExecuteSqlCommand(sQuery);
+                    }
+                    else
+                    {
+                        RecordFrm.txtBillNo.Text = GenerateBillNo();
+                    }
 
                     sQuery = "insert into billing values (:1,:2,:3,:4,:5,:6,:7)";
                     db.Database.ExecuteSqlCommand(sQuery,
@@ -217,7 +202,7 @@ namespace Modules.Billing
                     RecordFrm.btnSave.Enabled = false;
                     RecordFrm.btnCancel.Text = "Exit";
 
-                
+                }
             }
             catch (Exception ex)
             {
